% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{camera_orbit}
\alias{camera_orbit}
\title{Orbit a camera around an axis}
\usage{
camera_orbit(camera, axis = c(0, 0, 1), angle_degrees)
}
\arguments{
\item{camera}{A camera list from \code{camera()} or \code{camera_auto()}.}

\item{axis}{Rotation axis as a length-3 vector. Default \code{c(0, 0, 1)} (Z axis).}

\item{angle_degrees}{Rotation angle in degrees.}
}
\value{
A camera list with S3 class \code{"scimesh_camera"}.
}
\description{
Rotates a camera's eye position and up vector around its center
by a given angle about a rotation axis.  Useful for generating
turntable-style frame sequences.
}
\examples{
mesh <- generate_torus(c(0, 0, 0))
cam <- camera_auto(mesh, direction = c(1, 1, 1))
cam2 <- camera_orbit(cam, axis = c(0, 0, 1), angle_degrees = 90)

}
