% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{check_alpha}
\alias{check_alpha}
\title{Validate a Catmull-Rom parameterization exponent}
\usage{
check_alpha(alpha)
}
\arguments{
\item{alpha}{Requested exponent.}
}
\value{
The value as a double.
}
\description{
Validate a Catmull-Rom parameterization exponent
}
\keyword{internal}
