% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{check_flag}
\alias{check_flag}
\title{Validate a logical flag}
\usage{
check_flag(value, arg_name)
}
\arguments{
\item{value}{Requested value.}

\item{arg_name}{Name of the argument, used in error messages.}
}
\value{
A single logical.
}
\description{
Validate a logical flag
}
\keyword{internal}
