% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline.R
\name{check_sample_count}
\alias{check_sample_count}
\title{Validate a sample count}
\usage{
check_sample_count(samples_per_segment)
}
\arguments{
\item{samples_per_segment}{Requested number of samples.}
}
\value{
The value as an integer.
}
\description{
Validate a sample count
}
\keyword{internal}
