% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{default_font}
\alias{default_font}
\title{Path of the font used for text labels}
\usage{
default_font()
}
\value{
A character scalar: the path to an existing font file.
}
\description{
Text labels are drawn with a TrueType font that ships with the package
(\code{inst/extdata/Inter-Regular.ttf}, SIL Open Font License 1.1), so labels
look the same everywhere and no system font is required.  This function
returns the path of the font that will be used by default.
}
\details{
It can be overridden without touching any code by setting the
\code{SCIMESH_FONT} environment variable to the path of another \code{.ttf}
file, or per call by passing \code{font_file} to
\code{\link{text_layer}()}, \code{\link{text_extent}()} and friends.
}
\examples{
default_font()

}
\seealso{
\code{\link{font_info}}, \code{\link{text_layer}}
}
