% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{generate_axes}
\alias{generate_axes}
\title{Generate XYZ axis arrows as cylinder meshes}

\description{
Creates three coloured arrow meshes (red X, green Y, blue Z) from
a centre point.
}
\usage{
generate_axes(center = c(0, 0, 0), size = 1, shaft_radius = 0.02)
}
\arguments{
\item{center}{Length-3 vector: origin of the axes.}

\item{size}{Length of each axis.}

\item{shaft_radius}{Cylinder radius for axis shafts.}
}
\value{
A mesh descriptor list suitable for \code{render_mesh()}
  or inclusion in a scene list.
}

\examples{
axes_mesh <- generate_axes(size = 2)
nrow(axes_mesh$vertices)

}
