% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{generate_bbox}
\alias{generate_bbox}
\title{Generate a wireframe bounding box mesh}

\description{
Creates 12 edge segments around an axis-aligned bounding box.
}
\usage{
generate_bbox(bbox, color = c(0, 0, 0, 1), radius = 0.01)
}
\arguments{
\item{bbox}{A bounding box list from \code{mesh_bbox()}, or a mesh
descriptor (in which case \code{mesh_bbox()} is called).}

\item{color}{RGBA colour for the edges (length 4, 0-1 scale).}

\item{radius}{Cylinder radius for the edges.}
}
\value{
A mesh descriptor list suitable for \code{render_mesh()}
  or inclusion in a scene list.
}

\examples{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 2, 3))
bbox_mesh <- generate_bbox(mesh)
nrow(bbox_mesh$vertices)

}
