% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{generate_tetrahedron}
\alias{generate_tetrahedron}
\title{Generate a tetrahedron mesh}
\usage{
generate_tetrahedron(p0, p1, p2, p3, color = c(0.7, 0.7, 0.7, 1))
}
\arguments{
\item{p0}{Length-3 vector: first vertex.}

\item{p1}{Length-3 vector: second vertex.}

\item{p2}{Length-3 vector: third vertex.}

\item{p3}{Length-3 vector: fourth vertex.}

\item{color}{Length-4 RGBA colour.}
}
\value{
A mesh descriptor list.
}
\description{
Creates a tetrahedron (triangular pyramid) from four arbitrary
3D points.
}
\examples{
mesh <- generate_tetrahedron(
  c(0, 0, 0), c(1, 0, 0),
  c(0.5, 1, 0), c(0.5, 0.5, 1))
nrow(mesh$vertices)

}
