% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_export.R
\name{image_crop}
\alias{image_crop}
\title{Crop an image to a rectangular region}
\usage{
image_crop(image, x, y, w, h)
}
\arguments{
\item{image}{An image list returned by \code{render_mesh()} or similar.}

\item{x}{Left edge of the crop region (0-based pixel coordinate).}

\item{y}{Top edge of the crop region (0-based pixel coordinate).}

\item{w}{Crop width in pixels.}

\item{h}{Crop height in pixels.}
}
\value{
A new image list with the cropped dimensions.
}
\description{
Crop an image to a rectangular region
}
\examples{
cube <- generate_cuboid(c(0, 0, 0), c(1, 1, 1))
img <- render_mesh(cube$vertices, cube$triangles)
img <- image_crop(img, 100, 50, 400, 300)

}
