% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_export.R
\name{image_grow}
\alias{image_grow}
\title{Grow an image by adding padding}
\usage{
image_grow(image, top, bottom, left, right, background)
}
\arguments{
\item{image}{An image list.}

\item{top}{Number of pixel rows to add above.}

\item{bottom}{Number of pixel rows to add below.}

\item{left}{Number of pixel columns to add to the left.}

\item{right}{Number of pixel columns to add to the right.}

\item{background}{Numeric vector of length 4 with RGBA values in
\code{[0, 1]}.}
}
\value{
A new image list with the expanded dimensions.
}
\description{
Expands the canvas by adding pixel rows/columns filled with a
background colour.
}
\examples{
cube <- generate_cuboid(c(0, 0, 0), c(1, 1, 1))
img <- render_mesh(cube$vertices, cube$triangles)
img <- image_grow(img, 10, 10, 20, 20, c(1, 1, 1, 1))

}
