% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_export.R
\name{image_rotate_90}
\alias{image_rotate_90}
\title{Rotate an image by 90 degrees}
\usage{
image_rotate_90(image, clockwise = TRUE)
}
\arguments{
\item{image}{An image list.}

\item{clockwise}{Logical, if \code{TRUE} (default) rotates clockwise,
otherwise counter-clockwise.}
}
\value{
A new image list with width and height swapped.
}
\description{
Rotate an image by 90 degrees
}
