% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_export.R
\name{image_scale}
\alias{image_scale}
\title{Scale an image (nearest-neighbour)}
\usage{
image_scale(image, new_width, new_height)
}
\arguments{
\item{image}{An image list.}

\item{new_width}{Target width in pixels.}

\item{new_height}{Target height in pixels.}
}
\value{
A new image list with the new dimensions.
}
\description{
Resizes the image to the given dimensions using nearest-neighbour
interpolation.
}
