% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{mesh_bbox}
\alias{mesh_bbox}
\title{Compute the axis-aligned bounding box of a mesh}

\description{
Compute the axis-aligned bounding box of a mesh
}
\usage{
mesh_bbox(mesh)
}
\arguments{
\item{mesh}{A mesh descriptor list with \code{vertices}.}
}
\value{
A list with \code{min} and \code{max} (each length-3 numeric).
}

\examples{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 2, 3))
bb <- mesh_bbox(mesh)
bb$min
bb$max

}
