% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{stack_horizontal}
\alias{stack_horizontal}
\title{Stack images horizontally}
\usage{
stack_horizontal(
  ...,
  colorbar = NULL,
  colorbar_height = 80L,
  colorbar_width = 80L,
  background = c(0, 0, 0, 0),
  colorbar_side = c("right", "left"),
  crop = FALSE
)
}
\arguments{
\item{...}{Images from \code{render_mesh()} or
\code{render_scene()}, or a list of images.}

\item{colorbar}{Optional colorbar.}

\item{colorbar_height}{Height of the colorbar in pixels.}

\item{colorbar_width}{Width of the colorbar in pixels.}

\item{background}{Background RGBA color for padding.}

\item{colorbar_side}{Side for the colorbar: \code{"right"} (default)
or \code{"left"}.}

\item{crop}{If \code{TRUE}, crop whitespace from the output.}
}
\value{
A composed image list.
}
\description{
Stacks a list of rendered images side by side. A convenience
wrapper around \code{compose_layout()}.
}
\examples{
mesh1 <- generate_cuboid(c(-2, 0, 0), c(0.5, 1, 1), c(1, 0, 0, 1))
mesh2 <- generate_cuboid(c( 2, 0, 0), c(0.5, 1, 1), c(0, 0, 1, 1))
img1 <- render_mesh(mesh1$vertices, mesh1$triangles)
img2 <- render_mesh(mesh2$vertices, mesh2$triangles)
result <- stack_horizontal(img1, img2)
tmp_file <- tempfile(fileext = ".png")
write_png(result, tmp_file)

}
