% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{translate_mesh}
\alias{translate_mesh}
\title{Translate a mesh}
\usage{
translate_mesh(mesh, translation)
}
\arguments{
\item{mesh}{A mesh descriptor list.}

\item{translation}{Length-3 numeric vector (x, y, z).}
}
\value{
A new mesh descriptor list with translated vertices.
}
\description{
Vertex colors and normals are untouched: a translation does not change
the orientation of a surface.
}
\examples{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 1, 1))
moved <- translate_mesh(mesh, c(5, 0, 0))
colMeans(moved$vertices)

}
