% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{viridis_colormap}
\alias{viridis_colormap}
\title{Viridis colormap}

\description{
Returns the viridis color palette. A convenience wrapper around
\code{grDevices::hcl.colors} that mimics the viridis color scheme
without requiring extra packages.
}
\usage{
viridis_colormap(n, alpha = 1, direction = 1)
}
\arguments{
\item{n}{Number of colors.}

\item{alpha}{Alpha channel value (0-1).}

\item{direction}{Forward (1) or reversed (-1) direction.}
}
\value{
A character vector of hex color strings.
}

\examples{
cols <- viridis_colormap(10)
plot(1:10, pch = 19, col = cols, cex = 3)

}
