% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorbar.R
\name{diverging_colormap}
\alias{diverging_colormap}
\title{Custom diverging colormap for neuroimaging}
\usage{
diverging_colormap(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
A character vector of hex color strings.
}
\description{
Returns a blue-white-red diverging colormap suitable for displaying
signed morphometry data (e.g. cortical thickness Z-scores).
}
\examples{
cols <- diverging_colormap(256)
plot(1:256, pch = 15, col = cols, cex = 2, axes = FALSE, xlab = "", ylab = "")

}
