% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{generate_cuboid}
\alias{generate_cuboid}
\title{Generate a cuboid mesh}

\description{
Creates an axis-aligned cuboid (box) centred at \code{center}
with the given half-extents along each axis.
}
\usage{
generate_cuboid(center, half_extents, color = c(0.7, 0.7, 0.7, 1))
}
\arguments{
\item{center}{Length-3 vector: centre of the cuboid.}

\item{half_extents}{Length-3 vector: half-width, half-height,
half-depth.}

\item{color}{Length-4 RGBA colour (0-1 scale).}
}
\value{
A mesh descriptor list.
}

\examples{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 2, 0.5))
nrow(mesh$vertices)
nrow(mesh$triangles)

}
