% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{generate_tetrahedron}
\alias{generate_tetrahedron}
\title{Generate a tetrahedron mesh}
\usage{
generate_tetrahedron(p0, p1, p2, p3, color = c(0.7, 0.7, 0.7, 1))
}
\arguments{
\item{p0, }{p1, p2, p3 Length-3 vectors: the four vertices.}

\item{color}{Length-4 RGBA colour.}
}
\value{
A mesh descriptor list.
}
\description{
Creates a tetrahedron (triangular pyramid) from four arbitrary
3D points.
}
