% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{rotate_mesh}
\alias{rotate_mesh}
\title{Rotate a mesh around an axis}
\usage{
rotate_mesh(mesh, angle_rad, axis = c(0, 0, 1))
}
\arguments{
\item{mesh}{A mesh descriptor list.}

\item{angle_rad}{Rotation angle in radians.}

\item{axis}{Length-3 numeric vector defining the rotation axis.}
}
\value{
A new mesh descriptor list with rotated vertices.
}
\description{
Rotate a mesh around an axis
}
