% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh_utils.R
\name{scale_mesh}
\alias{scale_mesh}
\title{Scale a mesh uniformly}
\usage{
scale_mesh(mesh, scale)
}
\arguments{
\item{mesh}{A mesh descriptor list.}

\item{scale}{A single numeric scale factor.}
}
\value{
A new mesh descriptor list with scaled vertices.
}
\description{
Scale a mesh uniformly
}
\examples{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 1, 1))
big <- scale_mesh(mesh, 3)
range(big$vertices[, 1])

}
