% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{stack_horizontal}
\alias{stack_horizontal}
\title{Stack images horizontally}
\usage{
stack_horizontal(
  ...,
  colorbar = NULL,
  colorbar_height = 80L,
  colorbar_width = 80L,
  background = c(0, 0, 0, 0),
  crop = FALSE
)
}
\arguments{
\item{...}{Images from \code{render_mesh()} or
\code{render_scene()}, or a list of images.}

\item{colorbar}{Optional colorbar.}

\item{background}{Background RGBA color for padding.}
}
\value{
A composed image list.
}
\description{
Stacks a list of rendered images side by side. A convenience
wrapper around \code{compose_layout()}.
}
