% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{stack_vertical}
\alias{stack_vertical}
\title{Stack images vertically}

\description{
Stacks a list of rendered images vertically (one below another).
A convenience wrapper around \code{compose_layout()}.
}
\usage{
stack_vertical(
  ...,
  colorbar = NULL,
  colorbar_height = 80L,
  colorbar_width = 80L,
  background = c(0, 0, 0, 0),
  colorbar_side = c("right", "left"),
  crop = FALSE
)
}
\arguments{
\item{...}{Images from \code{render_mesh()} or
\code{render_scene()}, or a list of images.}

\item{colorbar}{Optional colorbar from \code{colorbar_horizontal()}
or \code{colorbar_vertical()}.}

\item{background}{Background RGBA color for padding.}
}
\value{
A composed image list.
}

\examples{
mesh1 <- generate_cuboid(c(0, 2, 0), c(1, 0.5, 1), c(1, 0, 0, 1))
mesh2 <- generate_cuboid(c(0, -2, 0), c(1, 0.5, 1), c(0, 1, 0, 1))
img1 <- render_mesh(mesh1$vertices, mesh1$triangles)
img2 <- render_mesh(mesh2$vertices, mesh2$triangles)
result <- stack_vertical(img1, img2)
\dontrun{write_png(result, "stacked_v.png")}

}
