% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_export.R
\name{write_png}
\alias{write_png}
\title{Write a rendered image to a PNG file}
\usage{
write_png(image, filename)
}
\arguments{
\item{image}{An image list returned by \code{render_mesh()} or
\code{render_scene()}.}

\item{filename}{Output PNG file path.}
}
\description{
Writes the output of \code{render_mesh()} or \code{render_scene()}
to a PNG file using the built-in C++ PNG writer (stb_image_write).
No additional R packages are required.
}
\examples{
\dontrun{
mesh <- generate_cuboid(c(0, 0, 0), c(1, 1, 1))
img <- render_mesh(mesh$vertices, mesh$triangles)
write_png(img, "cuboid.png")
}

}
